Extract UaSelectCard to its own file, byte-for-byte (#2324)
The pure move, so the repaint that follows has nothing structural in it.
`UaSelectCard` leaves `FactionSelectCard.tsx` for
`components/selectCard/UaSelectCard.tsx`, matching how `components/taskCard/`
is laid out and following the shape #2322 set for S.N.I.D.E. The manifest row
in `factions/ua.ts` is unchanged except that it now points at the new module
and drops the `.then((m) => ({ default: m.UaSelectCard }))` unwrap, because
the new file default-exports the way every task card does. The dispatcher's
three UA-only imports (`UaSigil`, `UaMandala`, the `uaAtoms` four) go with it.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` of
`<FactionSelectCard faction="ua" />` across all six state x members
combinations is byte-identical before and after this commit (50,428 bytes,
`diff` clean). The probe was a throwaway; the standing guards are
`factionSelectCardDispatch.test.tsx` and `__tests__/uaDesktopSkin.test.tsx`,
both green and both repointed at the new module.

THE DISPATCHER'S "the tiles are dark" IS NOT CARRIED IN. It was already
deleted from the dispatcher by #2322; UA is the counter-example that proved
it false, so the new file says so in as many words instead.

THE WORDMARK KEY IS UNTOUCHED. #2332 renamed `names.ua` to "Unwavering
Artisans", but this tile draws `feed:factionSelect.ua.wordmark`, a separate
key still reading "UA" — verified in the probe output, which holds six
`>UA<`, one per state x members combination.
