+++ frontend/src/__tests__/uaDesktopSkin.test.tsx
@@ -25,7 +25,7 @@ import type { CommentOut } from '../api/comments'
 import type { CharacterOut } from '../api/auth'
 import UaTaskCard from '../components/taskCard/UaTaskCard'
 import UaFactionHero from '../components/factionHero/UaFactionHero'
-import { UaSelectCard } from '../components/selectCard/FactionSelectCard'
+import UaSelectCard from '../components/selectCard/UaSelectCard'
 import UaFeedFrame from '../components/feed/UaFeedFrame'
 import UaAvatar from '../components/avatar/UaAvatar'
 import UaBackdrop from '../components/backdrop/UaBackdrop'
