+++ frontend/src/factions/ua.ts
@@ -29,7 +29,7 @@ const UaTaskDetail = lazyArchetype(() => import('../pages/taskDetail/archetypes/
 const UaVote = lazyArchetype(() => import('../components/vote/UaVote'))
 const UaPraxisCard = lazyArchetype(() => import('../components/praxisCard/desktop/UaPraxisCard'))
 const UaSigilAdapter = lazyArchetype(() => import('../components/sigil/FactionSigil').then((m) => ({ default: m.UaSigilAdapter })))
-const UaSelectCard = lazyArchetype(() => import('../components/selectCard/FactionSelectCard').then((m) => ({ default: m.UaSelectCard })))
+const UaSelectCard = lazyArchetype(() => import('../components/selectCard/UaSelectCard'))
 
 export const UA_MANIFEST: FactionManifest = {
   slug: 'ua',
